+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.2.0 (30/11/2023)
+
+### New Feature
+- Added support for multipart file upload
+- Print version number with -v flag
+
+### Improvements
+- If external response file is provided, response will always be stored in it
+
+### Fix
+- Close request payload file at the end
+
 ## v1.1.0 (24/10/2023)
 
 ### New Feature